+++ prisma/migrations/20240803163908_add_column_description_on_integrations_table/migration.sql
@@ -1,0 +1,8 @@
+-- AlterTable
+ALTER TABLE "Dify" ADD COLUMN     "description" VARCHAR(255);
+
+-- AlterTable
+ALTER TABLE "OpenaiBot" ADD COLUMN     "description" VARCHAR(255);
+
+-- AlterTable
+ALTER TABLE "Typebot" ADD COLUMN     "description" VARCHAR(255);