+++ src/api/integrations/typebot/dto/typebot.dto.ts
@@ -18,6 +18,7 @@ export class PrefilledVariables {
 
 export class TypebotDto {
   enabled?: boolean;
+  description?: string;
   url: string;
   typebot?: string;
   expire?: number;